Cleanup for rules.
