+++ .eslintrc.json
@@ -3,6 +3,7 @@
     "airbnb",
     "plugin:prettier/recommended"
   ],
+  "parser": "babel-eslint",
   "env": {
     "browser": true,
     "es6": true,
@@ -10,6 +11,10 @@
     "serviceworker": true
   },
   "rules": {
-    "react/jsx-filename-extension": [1, { "extensions": [".js"] }]
+    "react/jsx-filename-extension": [1, { "extensions": [".js"] }],
+    "jsx-a11y/anchor-is-valid": [ "error", {
+      "aspects": [ "noHref" ]
+    }],
+    "no-param-reassign": [2, { "props": false }]
   }
 }
