+++ src/components/05_pages/Permissions/Permissions.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { css } from 'emotion';
 
+const styles = {
+  title: css`
+    text-decoration: underline;
+  `,
+};
+
 const Permissions = () => (
   <div>
     <h1 className={styles.title}>Permissions</h1>
     <p>This will be the permissions page.</p>
   </div>
 );
-
-const styles = {
-  title: css`
-    text-decoration: underline;
-  `,
-};
 
 export default Permissions;
