+++ src/routes.js
@@ -1,0 +1,8 @@
+import Permissions from './components/05_pages/Permissions/Permissions';
+
+// @todo Share this with Drupal
+const routes = {
+  '/admin/people/permissions': Permissions,
+};
+
+export default routes;
